requirements.txt
streamlit
pandas
altair
openai
python-dotenv
